+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
-Adds full support for the [Customers V3 API](https://developer.bigcommerce.com/api-reference/customer-subscribers/v3-customers-api)
+Adds better support for the [Customers V3 API](https://developer.bigcommerce.com/api-reference/customer-subscribers/v3-customers-api)
 
 Including:
-- Customers API
-- Customer Addresses API
+- Customers
+- Customer Addresses
+- Customer Attributes
+